+++ qupath-core/src/main/java/qupath/lib/projects/DefaultProject.java
@@ -547,14 +547,14 @@ class DefaultProject implements Project<BufferedImage> {
 		}
 		
 		@Override
-		public Collection<URI> getServerURIs() throws IOException {
+		public Collection<URI> getURIs() throws IOException {
 			if (serverBuilder == null)
 				return Collections.emptyList();
 			return serverBuilder.getURIs();
 		}
 		
 		@Override
-		public boolean updateServerURIs(Map<URI, URI> replacements) throws IOException {
+		public boolean updateURIs(Map<URI, URI> replacements) throws IOException {
 			var builderBefore = serverBuilder;
 			serverBuilder = serverBuilder.updateURIs(replacements);
 			boolean changes = builderBefore != serverBuilder;